refactor(ledger): give the fee surface the per-resource route convention :hammer:
The fee and billing HTTP surface was organized by LAYER while every other
family in http/in is organized by RESOURCE: fees_v2_handler.go held all twelve
transport shells, fees_v2_register.go all twelve Huma registrars, and
fees_routes.go a shared guard table feeding them. huma_contract_mount.go showed
it — one line in a list of twenty took four handlers instead of one.

Fees now follow the CRM/ledger pattern exactly. Four resources, three files
each, and one Register<R>V2RoutesToApp per resource on the mount:

  fee_package_{core,handler,routes}.go
  fee_estimate_{core,handler,routes}.go
  billing_package_{core,handler,routes}.go
  billing_calculate_{core,handler,routes}.go

fee_ledger_scope.go keeps what is genuinely shared: the ledger-scope guards and
FeeV2Path/parseFeeV2Path.

Dissolved because the per-resource pattern supplies them: feeGuardRoutes,
attachFeeGuards, feeChainPath, feeSpecPathParam, feeBasePathV2, and
protectedFees (byte-identical to protectedMidaz). feeOpSuffixV2 gives way to the
shared v2OpSuffix — same "V2" value, so every operation ID is unchanged, and CRM
is the precedent for a v2-only family using the shared suffix.

Two guards had to move rather than go:

- feeChainPath derived the Fiber path from the OpenAPI one, so the two could not
  drift. TestFeesV2RoutesParameterNamesAgree already reads the parameter names
  off the live Huma document and compares them to Fiber's, so drift still fails
  the build; each routes file now writes both spellings, as holder_routes.go
  does.
- fees_authz_slug_test.go AST-parsed protectedFees out of fees_routes.go BY
  FILENAME. It is now TestAuthz_FeeResources_AuthorizeUnderMidazAppName, which
  drives all twelve production routes through a capturing authz server and
  asserts the forwarded product is midaz — the runtime forward rather than a
  literal in source, and it covers the two single-verb compute endpoints the
  five-verb table in routes_test.go cannot express.

Also renamed the transaction v2 files to suffix the version instead of
infixing it (transaction_v2_output.go -> transaction_output_v2.go), so each v2
file sorts beside its v1 sibling, and folded
TestUnifiedContract_NoDuplicateOperationIDs into contract_operation_ids_test.go:
it reads the live document where its neighbour reads the committed dump, and
v2_register_shared_test.go named a production file that no longer exists.

route_table.golden loses eight lines. The fee guard chain and Huma terminal now
register adjacently per resource, so Fiber merges them into one route row (3 + 1
becomes 4) as it already did for every other resource; no path is added or
removed. routeTableMinRows follows to one below the new count.

Paths, operation IDs, authz tuples and route order are unchanged. Pointers in
permissions.yaml, RBAC-NAMESPACES, the X1 RBAC runbook, STRUCTURE and the llms
manifests follow the renames.
