+++ components/ledger/internal/adapters/http/in/fees_authz_slug_test.go
@@ -5,134 +5,93 @@
 package in
 
 import (
-	"go/ast"
-	"go/parser"
-	"go/token"
+	"net/http/httptest"
 	"testing"
 
+	"github.com/LerianStudio/lib-auth/v3/auth/middleware"
+	"github.com/gofiber/fiber/v3"
+	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	ledgerMiddleware "github.com/LerianStudio/midaz/v4/components/ledger/internal/adapters/http/in/middleware"
+	pkgHTTP "github.com/LerianStudio/midaz/v4/pkg/net/http"
 )
 
-// parseFeesRoutesFile parses fees_routes.go into an AST. Tests run with the
-// package directory as the working directory, so the source file is reachable by
-// its base name.
-func parseFeesRoutesFile(t *testing.T) *ast.File {
-	t.Helper()
+// TestAuthz_FeeResources_AuthorizeUnderMidazAppName pins the authz application slug of
+// every fee and billing route to the ledger core's own "midaz" slug (midazName). Fee and
+// billing are a product embedded in the ledger V4 binary, and BOLA "one identity, one
+// slug" in the declaration receiver (plugin-identity :4001) forbids a route serving a
+// second slug behind the same identity.
+//
+// It is TestAuthz_RoutingResources_AuthorizeUnderMidazAppName for the fee surface: the
+// four PRODUCTION registrars are driven through a capturing authz server that records
+// the forwarded product, so a registrar repointed at a fees-specific slug makes the
+// recorded product diverge. All twelve ops are swept, including the two single-verb
+// compute endpoints the five-verb table in routes_test.go cannot express.
+//
+// The capturing server always denies (authorized=false). The auth middleware forwards
+// the product to the authz service BEFORE reading the decision, so the product is
+// captured on every request while the 403 short-circuits the chain — the business
+// terminals never run, so zero-value handlers are sufficient.
+//
+// NOT parallel: libProblem.Install swaps a process-global huma.NewError hook and Huma
+// validation uses process-global sync.Pools; concurrent builds cross-contaminate.
+func TestAuthz_FeeResources_AuthorizeUnderMidazAppName(t *testing.T) {
+	orgID := uuid.New()
+	ledgerID := uuid.New()
+	resourceID := uuid.New()
+	base := "/v2/organizations/" + orgID.String() + "/ledgers/" + ledgerID.String()
 
-	fset := token.NewFileSet()
+	var product string
 
-	file, err := parser.ParseFile(fset, "fees_routes.go", nil, parser.ParseComments)
-	require.NoError(t, err, "fees_routes.go must parse")
+	srv := newAuthzProductCapture(t, &product)
+	defer srv.Close()
 
-	return file
-}
+	auth := &middleware.AuthClient{Address: srv.URL, Enabled: true}
 
-// TestProtectedFeesAuthorizesUnderMidazSlug pins the authz application slug of every
-// fee and billing route to the ledger core's own "midaz" slug (midazName), the single
-// literal shared with protectedMidaz. Fee/billing is a product embedded in the ledger
-// V4 binary, and BOLA "one identity, one slug" in the declaration receiver
-// (plugin-identity :4001) forbids a route serving a second slug behind the same
-// identity. The invariant has no runtime seam without a full auth server, so it is
-// pinned at the definition site: protectedFees must call auth.Authorize with midazName.
-func TestProtectedFeesAuthorizesUnderMidazSlug(t *testing.T) {
-	t.Parallel()
+	f := fiber.New(fiber.Config{ErrorHandler: pkgHTTP.CanonicalFiberErrorHandler})
 
-	file := parseFeesRoutesFile(t)
+	// Mirror production: the ledger registers ErrorEnvelope on the app root.
+	f.Use(ledgerMiddleware.ErrorEnvelope())
 
-	var (
-		found     bool
-		firstArg  string
-		inspected bool
-	)
+	mountFeesV2Routes(f, auth, nil)
 
-	ast.Inspect(file, func(n ast.Node) bool {
-		fn, ok := n.(*ast.FuncDecl)
-		if !ok || fn.Name.Name != "protectedFees" {
-			return true
-		}
+	token := "Bearer " + guardBearerToken(t)
 
-		found = true
-
-		ast.Inspect(fn.Body, func(inner ast.Node) bool {
-			call, ok := inner.(*ast.CallExpr)
-			if !ok {
-				return true
-			}
-
-			sel, ok := call.Fun.(*ast.SelectorExpr)
-			if !ok || sel.Sel.Name != "Authorize" || len(call.Args) == 0 {
-				return true
-			}
-
-			inspected = true
-
-			if ident, ok := call.Args[0].(*ast.Ident); ok {
-				firstArg = ident.Name
-			}
-
-			return false
-		})
-
-		return false
-	})
-
-	require.True(t, found, "protectedFees must exist in fees_routes.go")
-	require.True(t, inspected, "protectedFees must call auth.Authorize")
-	assert.Equal(t, "midazName", firstArg,
-		"protectedFees must authorize under the shared midazName slug, not a fees-specific one")
-}
-
-// TestFeesApplicationNameConstRemoved asserts the fees-specific slug constant is gone,
-// leaving midazName (declared in routes.go) as the single source of the "midaz" literal
-// for the package. A reintroduced fees const would resurrect the retired two-slug model.
-func TestFeesApplicationNameConstRemoved(t *testing.T) {
-	t.Parallel()
-
-	file := parseFeesRoutesFile(t)
-
-	for _, decl := range file.Decls {
-		gen, ok := decl.(*ast.GenDecl)
-		if !ok || gen.Tok != token.CONST {
-			continue
-		}
-
-		for _, spec := range gen.Specs {
-			vs, ok := spec.(*ast.ValueSpec)
-			if !ok {
-				continue
-			}
-
-			for _, name := range vs.Names {
-				assert.NotEqual(t, "feesApplicationName", name.Name,
-					"the fees-specific slug constant must be removed; use midazName")
-			}
-		}
+	ops := []struct {
+		method string
+		path   string
+	}{
+		{fiber.MethodPost, base + "/packages"},
+		{fiber.MethodGet, base + "/packages"},
+		{fiber.MethodGet, base + "/packages/" + resourceID.String()},
+		{fiber.MethodPatch, base + "/packages/" + resourceID.String()},
+		{fiber.MethodDelete, base + "/packages/" + resourceID.String()},
+		{fiber.MethodPost, base + "/estimates"},
+		{fiber.MethodPost, base + "/billing-packages"},
+		{fiber.MethodGet, base + "/billing-packages"},
+		{fiber.MethodGet, base + "/billing-packages/" + resourceID.String()},
+		{fiber.MethodPatch, base + "/billing-packages/" + resourceID.String()},
+		{fiber.MethodDelete, base + "/billing-packages/" + resourceID.String()},
+		{fiber.MethodPost, base + "/billing/calculate"},
 	}
-}
 
-// TestFeesRoutesHasNoRetiredSlug asserts the retired fee-plugin slug appears nowhere in
-// fees_routes.go — neither as a string literal nor inside a comment. The forbidden token
-// is assembled from fragments so the source of this guard does not itself carry the very
-// literal a directory-wide grep is meant to find zero of.
-func TestFeesRoutesHasNoRetiredSlug(t *testing.T) {
-	t.Parallel()
+	require.Len(t, ops, len(feesV2FullRoutes), "the sweep must cover every mounted fee route")
 
-	retiredSlug := "plugin" + "-" + "fees"
+	for _, op := range ops {
+		t.Run(op.method+" "+op.path, func(t *testing.T) {
+			product = ""
 
-	file := parseFeesRoutesFile(t)
+			req := httptest.NewRequest(op.method, op.path, nil)
+			req.Header.Set("Authorization", token)
 
-	ast.Inspect(file, func(n ast.Node) bool {
-		if lit, ok := n.(*ast.BasicLit); ok && lit.Kind == token.STRING {
-			assert.NotContains(t, lit.Value, retiredSlug,
-				"no string literal in fees_routes.go may contain the retired fee-plugin slug")
-		}
-
-		return true
-	})
-
-	for _, group := range file.Comments {
-		assert.NotContains(t, group.Text(), retiredSlug,
-			"no comment in fees_routes.go may reference the retired fee-plugin slug")
+			resp, err := f.Test(req, fiber.TestConfig{Timeout: 0})
+			require.NoError(t, err)
+			require.Equalf(t, fiber.StatusForbidden, resp.StatusCode,
+				"%s %s must reach auth and be denied", op.method, op.path)
+			assert.Equalf(t, midazName, product,
+				"%s %s must authorize under the midaz appName, got %q", op.method, op.path, product)
+		})
 	}
 }
