+++ components/ledger/internal/adapters/http/in/huma_contract_mount.go
@@ -208,7 +208,10 @@ func (d HumaMountDeps) MountV2(group fiber.Router, api huma.API) {
 	RegisterInstrumentV2RoutesToApp(group, api, d.Auth, d.Instrument, d.CRMOptions)
 	RegisterEncryptionV2RoutesToApp(group, api, d.Auth, d.Encryption, d.CRMOptions)
 	RegisterAuditV2RoutesToApp(group, api, d.Auth, d.Audit, d.CRMOptions)
-	RegisterFeesV2RoutesToApp(group, api, d.Auth, d.FeePackage, d.Fee, d.BillingPackage, d.BillingCalculate, d.FeesOptions)
+	RegisterPackageV2RoutesToApp(group, api, d.Auth, d.FeePackage, d.FeesOptions)
+	RegisterFeeEstimateV2RoutesToApp(group, api, d.Auth, d.Fee, d.FeesOptions)
+	RegisterBillingPackageV2RoutesToApp(group, api, d.Auth, d.BillingPackage, d.FeesOptions)
+	RegisterBillingCalculateV2RoutesToApp(group, api, d.Auth, d.BillingCalculate, d.FeesOptions)
 	RegisterCompositionV2RoutesToApp(group, api, d.Auth, d.Composition, d.CompositionOptions)
 	RegisterOperationRouteV2RoutesToApp(group, api, d.Auth, d.OperationRoute, d.TransactionOptions)
 	RegisterTransactionRouteV2RoutesToApp(group, api, d.Auth, d.TransactionRoute, d.TransactionOptions)
