+++ components/ledger/internal/adapters/http/in/transaction_field_removal_v2_test.go
