+++ components/ledger/internal/adapters/http/in/transaction_mirror_reads_v2_integration_test.go
@@ -27,7 +27,7 @@ import (
 
 // This file is the happy-path behavioral proof for the three /v2 transaction mirror ops that carry
 // no dedicated create/lifecycle wire shape — GetTransactionV2, GetAllTransactionsV2, and
-// UpdateTransactionV2 (transaction_v2_mirror_handler.go). The unit suite only reaches these on
+// UpdateTransactionV2 (transaction_mirror_handler_v2.go). The unit suite only reaches these on
 // the invalid-path-ID guard branch; here they run against a LIVE app backed by testcontainers, over
 // a transaction seeded through a v1 create so the seed path is unaffected by the v2 field removals.
 //
