+++ components/ledger/internal/adapters/http/in/transaction_output_v2.go
