+++ components/ledger/internal/adapters/http/in/transaction_response_shape_v2_integration_test.go
