+++ components/ledger/internal/bootstrap/testdata/route_table.golden
@@ -37,11 +37,9 @@ DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/account-types/:id	4
 DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/accounts/:id	4
 DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/assets/:id	4
 DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/balances/:balance_id	4
-DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/billing-packages/:id	3
-DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/billing-packages/:id	1
+DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/billing-packages/:id	4
 DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/operation-routes/:operation_route_id	4
-DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/packages/:id	3
-DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/packages/:id	1
+DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/packages/:id	4
 DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/portfolios/:id	4
 DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/segments/:id	4
 DELETE	/v2/organizations/:organization_id/ledgers/:ledger_id/transaction-routes/:transaction_route_id	4
@@ -239,11 +237,9 @@ PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/account-types/:id	4
 PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/accounts/:id	4
 PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/assets/:id	4
 PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/balances/:balance_id	4
-PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/billing-packages/:id	3
-PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/billing-packages/:id	1
+PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/billing-packages/:id	4
 PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/operation-routes/:operation_route_id	4
-PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/packages/:id	3
-PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/packages/:id	1
+PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/packages/:id	4
 PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/portfolios/:id	4
 PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/segments/:id	4
 PATCH	/v2/organizations/:organization_id/ledgers/:ledger_id/settings	3
@@ -289,16 +285,12 @@ POST	/v2/organizations/:organization_id/ledgers/:ledger_id/account-types	4
 POST	/v2/organizations/:organization_id/ledgers/:ledger_id/accounts	4
 POST	/v2/organizations/:organization_id/ledgers/:ledger_id/accounts/:account_id/balances	4
 POST	/v2/organizations/:organization_id/ledgers/:ledger_id/assets	4
-POST	/v2/organizations/:organization_id/ledgers/:ledger_id/billing-packages	3
-POST	/v2/organizations/:organization_id/ledgers/:ledger_id/billing-packages	1
-POST	/v2/organizations/:organization_id/ledgers/:ledger_id/billing/calculate	3
-POST	/v2/organizations/:organization_id/ledgers/:ledger_id/billing/calculate	1
-POST	/v2/organizations/:organization_id/ledgers/:ledger_id/estimates	3
-POST	/v2/organizations/:organization_id/ledgers/:ledger_id/estimates	1
+POST	/v2/organizations/:organization_id/ledgers/:ledger_id/billing-packages	4
+POST	/v2/organizations/:organization_id/ledgers/:ledger_id/billing/calculate	4
+POST	/v2/organizations/:organization_id/ledgers/:ledger_id/estimates	4
 POST	/v2/organizations/:organization_id/ledgers/:ledger_id/holders/:id/accounts	4
 POST	/v2/organizations/:organization_id/ledgers/:ledger_id/operation-routes	4
-POST	/v2/organizations/:organization_id/ledgers/:ledger_id/packages	3
-POST	/v2/organizations/:organization_id/ledgers/:ledger_id/packages	1
+POST	/v2/organizations/:organization_id/ledgers/:ledger_id/packages	4
 POST	/v2/organizations/:organization_id/ledgers/:ledger_id/portfolios	4
 POST	/v2/organizations/:organization_id/ledgers/:ledger_id/segments	4
 POST	/v2/organizations/:organization_id/ledgers/:ledger_id/transaction-routes	4
