+++ llms.txt
@@ -40,7 +40,7 @@ The three deploy units form one core banking platform in one monorepo:
 
 - [Ledger Component](components/ledger/): Unified deploy unit — onboarding + transaction + CRM + fees in one process (:3002)
   - [Bootstrap](components/ledger/internal/bootstrap/): Composition root — config, DI, server lifecycle, `initCRM`, fee wiring, multi-tenant middleware
-  - [HTTP Handlers](components/ledger/internal/adapters/http/in/): Fiber handlers, routes, middleware, `fees_routes.go` + `fees_v2_register.go`, CRM routes (`holder_routes.go`, `holder_accounts_routes.go`, `instrument_routes.go`, `encryption_routes.go`, `audit_routes.go`, `composition_routes.go`) with their `*_handler.go` transports and `*_core.go` cores
+  - [HTTP Handlers](components/ledger/internal/adapters/http/in/): Fiber handlers, routes, middleware, fee/billing routes (`fee_package_routes.go`, `fee_estimate_routes.go`, `billing_package_routes.go`, `billing_calculate_routes.go`), CRM routes (`holder_routes.go`, `holder_accounts_routes.go`, `instrument_routes.go`, `encryption_routes.go`, `audit_routes.go`, `composition_routes.go`) with their `*_handler.go` transports and `*_core.go` cores
   - [PostgreSQL Repos](components/ledger/internal/adapters/postgres/): Onboarding + transaction repositories
   - [MongoDB Repos](components/ledger/internal/adapters/mongodb/): Metadata + fees storage
   - [Services](components/ledger/internal/services/): CQRS command/query use cases + embedded `fees/`
